refactor: rewrite helper to ts
