+++ src/_modules/collapseWhitespace.mjs
@@ -1,4 +1,4 @@
-import { isComment } from '../helpers.mjs';
+import { isComment } from '../helpers';
 
 const noWhitespaceCollapseElements = new Set([
     'script',
