+++ src/_modules/minifySvg.mjs
@@ -1,4 +1,4 @@
-import { optionalImport } from '../helpers.mjs';
+import { optionalImport } from '../helpers';
 
 /** Minify SVG with SVGO */
 export default async function minifySvg(tree, options, svgoOptions = {}) {
