+++ src/_modules/minifyUrls.mjs
@@ -1,4 +1,4 @@
-import { optionalImport } from '../helpers.mjs';
+import { optionalImport } from '../helpers';
 
 // Adopts from https://github.com/kangax/html-minifier/blob/51ce10f4daedb1de483ffbcccecc41be1c873da2/src/htmlminifier.js#L209-L221
 const tagsHaveUriValuesForAttributes = new Set([
