+++ src/_modules/removeEmptyAttributes.mjs
@@ -1,4 +1,4 @@
-import { isEventHandler } from '../helpers.mjs';
+import { isEventHandler } from '../helpers';
 
 const safeToRemoveAttrs = {
     id: null,
